+++ src/commands/trading-wallet/CreateWalletCommand.js
@@ -53,7 +53,7 @@ class CreateWalletCommand extends CredentialBasedCommand {
     const personalWalletAddressArg = new CommandArg('personal-wallet-address',
       'string', 'eoa', 'The personal wallet address (EOA).', 1, true)
     const keystoreFilePathArg = new CommandArg('keystore-file-path',
-      'string', 'ksp', 'The private key file path.', 1, true)
+      'string', 'ksp', 'The keystore file path.', 1, true)
     const draftArg = new CommandArg('draft',
       'boolean', 'd', 'If set, it returns the TransactionObjectDraft.', 0, false, false)
     const rawTxArg = new CommandArg('raw-tx',