+++ src/commands/trading-wallet/WithdrawCommand.js
@@ -60,7 +60,7 @@ class WithdrawCommand extends CredentialBasedCommand {
     const tokenArg = new CommandArg('token',
       'string', 'tk', 'The token address.', 1, true)
     const keystoreFilePathArg = new CommandArg('keystore-file-path',
-      'string', 'ksp', 'The private key file path.', 1, true)
+      'string', 'ksp', 'The keystore file path.', 1, true)
     const draftArg = new CommandArg('draft',
       'boolean', 'd', 'If set, it returns the TransactionObjectDraft.', 0, false, false)
     const rawTxArg = new CommandArg('raw-tx',